# demo/openapi.json
{
  "openapi":"3.0.3",
  "info":{"title":"Demo Backend API","version":"1.0.0","description":"로그인 후 항목을 생성하고 응답 ID로 조회하는 시나리오를 시험하세요.\n\n```mermaid\nflowchart LR\n  login[로그인] --> create[항목 생성] --> read[항목 조회]\n```"},
  "servers":[{"url":"/demo-api"}],
...
  "paths":{
...
    "/session":{"get":{"tags":["인증"],"summary":"세션 확인","operationId":"session","parameters":[{"in":"cookie","name":"studio-session","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"브라우저 쿠키 인증","content":{"application/json":{"schema":{"type":"object","properties":{"authenticated":{"type":"boolean"}}}}}}}}},
...
  },
  "components":{"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer"}}}
}